+++ components/Registration/TextInput.tsx
@@ -1,0 +1,16 @@
+import { useField } from "formik";
+
+const TextInput = ({ label, ...props }) => {
+    const [field, meta] = useField(props.name);
+    return (
+      <div style={{marginBottom: "7px"}}>
+        <label htmlFor={props.name}>{label}</label>
+        {meta.touched && meta.error ? (
+          <div className="registration-error">{meta.error}</div>
+        ) : null}
+        <input {...field} {...props} />
+      </div>
+    );
+  };
+
+export default TextInput
