--- pages/registration/confirm.tsx
@@ -1,17 +1,0 @@
-import Image from "next/image";
-import logo from "/public/images/surma_logo.svg";
-
-export default function Confirm() {
-  return (
-    <div style={{ padding: "15px" }}>
-      <Image src={logo} width={50} height={50} alt="logo" />
-      <h2>Kiitos ilmoittautumisestasi!</h2>
-      <p>Saat pian sähköpostiisi automaattisen vahvistusviestin.</p>
-      <p>
-        Ennen turnauksen alkua saat viestin, jossa on ohjeet tunnusten luomiseen
-        turnausjärjestelmään. Lisäksi tuomaristo ottaa vielä yhteyttä ennen
-        turnauksen alkua.
-      </p>
-    </div>
-  );
-}
